% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{Example single-cell datasets}
\format{
A list of 4 elements:
\describe{
  \item{SS2}{Seurat object from Smart-seq2 platform}
  \item{tenx}{Seurat object from 10x Chromium platform}
  \item{meta_SS2}{data.frame, cell-level metadata for SS2}
  \item{meta_tenx}{data.frame, cell-level metadata for tenx}
}
}
\usage{
ExampleData
}
\description{
A list of Seurat objects and corresponding metadata from two 10x Genomics
peripheral blood samples (SS2 and tenx).
}
\examples{
data("ExampleData")
names(ExampleData)
head(ExampleData$meta_SS2)
}
\keyword{datasets}
